mfs: split into components [#1345]
Separate out share and xattrs support.
