+++ src/dosext/mfs/mfs.c
@@ -165,7 +165,6 @@ TODO:
 #include <sys/ioctl.h>
 #include <sys/stat.h>
 #include <sys/types.h>
-#include <sys/xattr.h>
 #include <unistd.h>
 #include <time.h>
 #include <errno.h>
@@ -176,12 +175,6 @@ TODO:
 #include <ctype.h>
 #include <stdint.h>	// types used for seek/size
 
-#if !DOSEMU
-#include <mach/message.h>
-#include <mach/exception.h>
-
-#include "dos.h"
-#else
 #include <dirent.h>
 #include <string.h>
 #ifdef HAVE_LIBBSD
@@ -190,7 +183,6 @@ TODO:
 #include <wctype.h>
 #include "emu.h"
 #include "int.h"
-#include "mfs.h"
 #include "lfn.h"
 #include "dos2linux.h"
 #include "doshelpers.h"
@@ -202,7 +194,9 @@ TODO:
 #include "utilities.h"
 #include "coopth.h"
 #include "lpt.h"
-#endif
+#include "share.h"
+#include "xattr.h"
+#include "mfs.h"
 
 #ifdef __linux__
 #include <linux/msdos_fs.h>
@@ -275,22 +269,6 @@ static int emufs_loaded = FALSE;
 enum {DRV_NOT_FOUND, DRV_FOUND, DRV_NOT_ASSOCIATED};
 
 enum { TYPE_NONE, TYPE_DISK, TYPE_PRINTER };
-struct file_fd
-{
-  char *name;
-  int idx;
-  int fd;
-  int type;
-  int dir_fd;
-  struct stat st;
-  int is_writable;
-  int read_allowed;
-  int share_mode;
-  u_short psp;
-  uint64_t seek;
-  uint64_t size;
-  int lock_cnt;
-};
 
 /* Need to know how many drives are redirected */
 static u_char redirected_drives = 0;
@@ -307,14 +285,10 @@ static void path_to_ufs(char *ufs, size_t ufs_offset, const char *path,
                         int PreserveEnvVar, int lowercase);
 static int dos_would_allow(char *fpath, const char *op, int equal);
 static void RemoveRedirection(int drive, cds_t cds);
-static int get_dos_xattr(const char *fname);
-static int set_dos_xattr(const char *fname, int attr);
-static int open_compat(int fd);
 
 static int drives_initialized = FALSE;
 
-#define MAX_OPENED_FILES 256
-static struct file_fd open_files[MAX_OPENED_FILES];
+struct file_fd open_files[MAX_OPENED_FILES];
 static int num_drives = 0;
 
 lol_t lol = 0;
@@ -386,514 +360,6 @@ static void lock_get(int fd, struct flock *fl)
     /* pretend nothing is locked */
     fl->l_type = F_UNLCK;
   }
-}
-
-static void downgrade_dir_lock(int dir_fd, int fd, off_t start)
-{
-    struct flock fl;
-    int err;
-
-    fl.l_type = F_RDLCK;
-    fl.l_start = start;
-    fl.l_len = 1;
-    /* set read lock and remove exclusive lock */
-    err = lock_set(dir_fd, &fl);
-    /* read lock should never fail (we put no write locks) */
-    if (err)
-        error("MFS: read lock failed, %s\n", strerror(errno));
-    flock(dir_fd, LOCK_UN);
-}
-
-static int do_open_dir(const char *dname, int *locked)
-{
-    int err;
-    int dir_fd = open(dname, O_RDONLY | O_DIRECTORY | O_CLOEXEC);
-    if (dir_fd == -1) {
-        error("MFS: failed to open %s: %s\n", dname, strerror(errno));
-        return -1;
-    }
-    *locked = 0;
-    /* lock directory (OFD) to avoid races */
-    err = flock(dir_fd, LOCK_EX);
-    if (err)
-        error("MFS: exclusive lock failed: %s\n", strerror(errno));
-    else
-        *locked = 1;
-    return dir_fd;
-}
-
-static struct file_fd *do_claim_fd(const char *name)
-{
-    int i;
-    struct file_fd *ret = NULL;
-
-    for (i = 0; i < MAX_OPENED_FILES; i++) {
-        struct file_fd *f = &open_files[i];
-        if (!f->name) {
-            f->name = strdup(name);
-            f->idx = i;
-            ret = f;
-            break;
-        }
-    }
-    if (!ret) {
-        error("MFS: too many open files\n");
-        leavedos(1);
-        return NULL;
-    }
-    ret->dir_fd = -1;
-    ret->seek = 0;
-    ret->size = 0;
-    return ret;
-}
-
-static struct file_fd *do_find_fd(const char *name)
-{
-    int i;
-    struct file_fd *ret = NULL;
-
-    for (i = 0; i < MAX_OPENED_FILES; i++) {
-        struct file_fd *f = &open_files[i];
-        if (f->name && strcmp(name, f->name) == 0) {
-            ret = f;
-            break;
-        }
-    }
-    return ret;
-}
-
-static int file_is_opened(int dir_fd, const char *name, int *r_err)
-{
-    struct stat st;
-    struct flock fl;
-    int err;
-
-    err = fstatat(dir_fd, name, &st, 0);
-    if (err) {
-        /* check for dangling symlink */
-        if (fstatat(dir_fd, name, &st, AT_SYMLINK_NOFOLLOW) == 0)
-            return 0;
-        *r_err = FILE_NOT_FOUND;
-        return -1;
-    }
-    fl.l_type = F_WRLCK;
-    fl.l_start = st.st_ino;
-    fl.l_len = 1;
-    lock_get(dir_fd, &fl);
-    return (fl.l_type == F_UNLCK ? 0 : 1);
-}
-
-static int do_mfs_creat(struct file_fd *f, const char *dname,
-        const char *fname, mode_t mode)
-{
-    int fd, dir_fd, err;
-    int locked = 0;
-
-    dir_fd = do_open_dir(dname, &locked);
-    if (dir_fd == -1)
-        return -1;
-    fd = openat(dir_fd, fname, O_RDWR | O_CLOEXEC | O_CREAT | O_TRUNC,
-            mode);
-    if (fd == -1)
-        goto err;
-    /* set compat mode */
-    err = open_compat(fd);
-    if (err)
-        goto err2;
-    err = fstat(fd, &f->st);
-    if (err)
-        goto err2;
-    if (locked)
-        downgrade_dir_lock(dir_fd, fd, f->st.st_ino);
-
-    f->fd = fd;
-    f->dir_fd = dir_fd;
-    f->share_mode = 0;
-    f->psp = sda_cur_psp(sda);
-    f->is_writable = 1;
-    f->read_allowed = 1;
-    return 0;
-
-err2:
-    unlinkat(dir_fd, fname, 0);
-    close(fd);
-err:
-    close(dir_fd);
-    return -1;
-}
-
-static struct file_fd *mfs_creat(char *name, mode_t mode)
-{
-    char *slash = strrchr(name, '/');
-    struct file_fd *f;
-    char *fname;
-    int err;
-    if (!slash)
-        return NULL;
-    f = do_claim_fd(name);
-    if (!f)
-        return NULL;
-    fname = slash + 1;
-    *slash = '\0';
-    err = do_mfs_creat(f, name, fname, mode);
-    *slash = '/';
-    if (err) {
-        free(f->name);
-        f->name = NULL;
-        return NULL;
-    }
-    return f;
-}
-
-enum { compat_lk_off = 0x100000000LL, noncompat_lk_off, denyR_lk_off,
-    denyW_lk_off, R_lk_off, W_lk_off };
-
-static int open_compat(int fd)
-{
-    struct flock fl;
-    fl.l_type = F_WRLCK;
-    fl.l_start = noncompat_lk_off;
-    fl.l_len = 1;
-    lock_get(fd, &fl);
-    if (fl.l_type != F_UNLCK)
-        return -1;
-    fl.l_type = F_RDLCK;
-    fl.l_start = compat_lk_off;
-    fl.l_len = 1;
-    return lock_set(fd, &fl);
-}
-
-static int open_share(int fd, int open_mode, int share_mode)
-{
-    int err;
-    struct flock fl;
-    int denyR = (share_mode == DENY_READ || share_mode == DENY_ALL);
-    int denyW = (share_mode == DENY_WRITE || share_mode == DENY_ALL);
-    /* inhibit compat mode */
-    fl.l_type = F_WRLCK;
-    fl.l_start = compat_lk_off;
-    fl.l_len = 1;
-    lock_get(fd, &fl);
-    if (fl.l_type != F_UNLCK)
-        return -1;
-    if (open_mode != O_WRONLY) {
-        /* read mode allowed? */
-        fl.l_type = F_WRLCK;
-        fl.l_start = denyR_lk_off;
-        fl.l_len = 1;
-        lock_get(fd, &fl);
-        if (fl.l_type != F_UNLCK)
-            return -1;
-    }
-    if (open_mode == O_WRONLY || open_mode == O_RDWR) {
-        /* write mode allowed? */
-        fl.l_type = F_WRLCK;
-        fl.l_start = denyW_lk_off;
-        fl.l_len = 1;
-        lock_get(fd, &fl);
-        if (fl.l_type != F_UNLCK)
-            return -1;
-    }
-    if (denyR) {
-        /* denyR allowed? */
-        fl.l_type = F_WRLCK;
-        fl.l_start = R_lk_off;
-        fl.l_len = 1;
-        lock_get(fd, &fl);
-        if (fl.l_type != F_UNLCK)
-            return -1;
-    }
-    if (denyW) {
-        /* denyW allowed? */
-        fl.l_type = F_WRLCK;
-        fl.l_start = W_lk_off;
-        fl.l_len = 1;
-        lock_get(fd, &fl);
-        if (fl.l_type != F_UNLCK)
-            return -1;
-    }
-
-    /* all checks passed, claim our locks */
-    fl.l_type = F_RDLCK;
-    fl.l_start = noncompat_lk_off;
-    fl.l_len = 1;
-    err = lock_set(fd, &fl);
-    if (err)
-        return err;
-    if (open_mode != O_WRONLY) {
-        fl.l_type = F_RDLCK;
-        fl.l_start = R_lk_off;
-        fl.l_len = 1;
-        err = lock_set(fd, &fl);
-        if (err)
-            return err;
-    }
-    if (open_mode == O_WRONLY || open_mode == O_RDWR) {
-        fl.l_type = F_RDLCK;
-        fl.l_start = W_lk_off;
-        fl.l_len = 1;
-        err = lock_set(fd, &fl);
-        if (err)
-            return err;
-    }
-    if (denyR) {
-        fl.l_type = F_RDLCK;
-        fl.l_start = denyR_lk_off;
-        fl.l_len = 1;
-        err = lock_set(fd, &fl);
-        if (err)
-            return err;
-    }
-    if (denyW) {
-        fl.l_type = F_RDLCK;
-        fl.l_start = denyW_lk_off;
-        fl.l_len = 1;
-        err = lock_set(fd, &fl);
-        if (err)
-            return err;
-    }
-    return 0;
-}
-
-static int file_is_ro(const char *fname, mode_t mode)
-{
-    int attr = get_dos_xattr(fname);
-    if (attr == -1)
-        return 0;
-    return !!(attr & READ_ONLY_FILE);
-}
-
-static int do_mfs_open(struct file_fd *f, const char *dname,
-        const char *fname, int flags, struct stat *st, int share_mode,
-        int *r_err)
-{
-    int fd, dir_fd, err;
-    int is_writable = (flags == O_WRONLY || flags == O_RDWR);
-    int is_readable = (flags == O_RDONLY || flags == O_RDWR);
-    /* XXX: force in READ mode, as needed by our share emulation w OFD locks */
-    int flags2 = (flags == O_WRONLY ? O_RDWR : flags);
-    int locked = 0;
-
-    *r_err = ACCESS_DENIED;
-    dir_fd = do_open_dir(dname, &locked);
-    if (dir_fd == -1)
-        return -1;
-    fd = openat(dir_fd, fname, flags2 | O_CLOEXEC);
-    if (fd == -1)
-        goto err;
-    if (!share_mode)
-        err = open_compat(fd);
-    else
-        err = open_share(fd, flags, share_mode);
-    if (err) {
-        *r_err = SHARING_VIOLATION;
-        goto err2;
-    }
-    if (locked)
-        downgrade_dir_lock(dir_fd, fd, st->st_ino);
-
-    f->st = *st;
-    f->fd = fd;
-    f->dir_fd = dir_fd;
-    f->share_mode = share_mode;
-    f->psp = sda_cur_psp(sda);
-    f->is_writable = is_writable;
-    f->read_allowed = is_readable;
-    return 0;
-
-err2:
-    close(fd);
-err:
-    close(dir_fd);
-    return -1;
-}
-
-static struct file_fd *mfs_open(char *name, int flags, struct stat *st,
-        int share_mode, int *r_err)
-{
-    char *slash = strrchr(name, '/');
-    struct file_fd *f;
-    char *fname;
-    int err;
-    if (!slash)
-        return NULL;
-    f = do_claim_fd(name);
-    if (!f)
-        return NULL;
-    fname = slash + 1;
-    *slash = '\0';
-    err = do_mfs_open(f, name, fname, flags, st, share_mode, r_err);
-    *slash = '/';
-    if (err) {
-        free(f->name);
-        f->name = NULL;
-        return NULL;
-    }
-    return f;
-}
-
-static int do_mfs_unlink(const char *dname, const char *fname, int force)
-{
-    int dir_fd, err, rc;
-    int locked = 0;
-
-    dir_fd = do_open_dir(dname, &locked);
-    if (dir_fd == -1)
-        return -1;
-    rc = file_is_opened(dir_fd, fname, &err);
-    switch (rc) {
-        case -1:
-            close(dir_fd);
-            return err;
-        case 0:
-            break;
-        case 1:
-            if (!force) {
-                close(dir_fd);
-                return ACCESS_DENIED;
-            }
-    }
-    err = unlinkat(dir_fd, fname, 0);
-    close(dir_fd);
-    if (err)
-        return FILE_NOT_FOUND;
-    return 0;
-}
-
-static int mfs_unlink(char *name)
-{
-    char *slash = strrchr(name, '/');
-    struct file_fd *f;
-    char *fname;
-    int err;
-    if (!slash)
-        return FILE_NOT_FOUND;
-    f = do_find_fd(name);
-    if (f && (f->share_mode || f->psp != sda_cur_psp(sda)))
-        return ACCESS_DENIED;
-    fname = slash + 1;
-    *slash = '\0';
-    err = do_mfs_unlink(name, fname, f && !f->share_mode &&
-        f->psp == sda_cur_psp(sda));
-    *slash = '/';
-    return err;
-}
-
-static int do_mfs_setattr(const char *dname, const char *fname,
-        const char *fullname, int attr, int force)
-{
-    int dir_fd, err, rc;
-    int locked = 0;
-
-    dir_fd = do_open_dir(dname, &locked);
-    if (dir_fd == -1)
-        return -1;
-    rc = file_is_opened(dir_fd, fname, &err);
-    switch (rc) {
-        case -1:
-            close(dir_fd);
-            return err;
-        case 0:
-            break;
-        case 1:
-            if (!force) {
-                close(dir_fd);
-                return ACCESS_DENIED;
-            }
-    }
-    err = set_dos_xattr(fullname, attr);
-    close(dir_fd);
-    return err;
-}
-
-static int mfs_setattr(char *name, int attr)
-{
-    char *slash = strrchr(name, '/');
-    struct file_fd *f;
-    char *fname;
-    char *fullname;
-    int err;
-    if (!slash)
-        return FILE_NOT_FOUND;
-    f = do_find_fd(name);
-    if (f && (f->share_mode || f->psp != sda_cur_psp(sda)))
-        return ACCESS_DENIED;
-    fname = slash + 1;
-    fullname = strdup(name);
-    *slash = '\0';
-    err = do_mfs_setattr(name, fname, fullname, attr, f && !f->share_mode &&
-        f->psp == sda_cur_psp(sda));
-    *slash = '/';
-    free(fullname);
-    return err;
-}
-
-static int do_mfs_rename(const char *dname, const char *fname,
-        const char *fname2, int force)
-{
-    int dir_fd, err, rc;
-    int locked = 0;
-
-    dir_fd = do_open_dir(dname, &locked);
-    if (dir_fd == -1)
-        return -1;
-    rc = file_is_opened(dir_fd, fname, &err);
-    switch (rc) {
-        case -1:
-            close(dir_fd);
-            return err;
-        case 0:
-            break;
-        case 1:
-            if (!force) {
-                close(dir_fd);
-                return ACCESS_DENIED;
-            }
-    }
-#ifdef HAVE_RENAMEAT2
-    err = renameat2(dir_fd, fname, -1, fname2, RENAME_NOREPLACE);
-    if (err && errno == EINVAL) {
-      error("MFS: RENAME_NOREPLACE unsupported here\n");
-      err = access(fname2, F_OK);
-      if (err)  // err means no file
-        err = renameat(dir_fd, fname, -1, fname2);
-    }
-#else
-    err = access(fname2, F_OK);
-    if (err)  // err means no file
-        err = renameat(dir_fd, fname, -1, fname2);
-#endif
-    close(dir_fd);
-    if (err)
-        return (errno == EEXIST ? ACCESS_DENIED : FILE_NOT_FOUND);
-    return 0;
-}
-
-static int mfs_rename(char *name, char *name2)
-{
-    char *slash = strrchr(name, '/');
-    struct file_fd *f;
-    char *fname;
-    int err;
-    if (!slash)
-        return FILE_NOT_FOUND;
-    f = do_find_fd(name);
-    if (f && (f->share_mode || f->psp != sda_cur_psp(sda)))
-        return ACCESS_DENIED;
-    fname = slash + 1;
-    *slash = '\0';
-    err = do_mfs_rename(name, fname, name2, f && !f->share_mode &&
-        f->psp == sda_cur_psp(sda));
-    *slash = '/';
-    return err;
-}
-
-static void mfs_close(struct file_fd *f)
-{
-    close(f->fd);
-    close(f->dir_fd);
-    free(f->name);
-    f->name = NULL;
 }
 
 static char *cds_flags_to_str(uint16_t flags) {
@@ -1224,92 +690,6 @@ static int fd_on_fat(int fd)
   return fstatfs(fd, &buf) == 0 && buf.f_type == MSDOS_SUPER_MAGIC;
 }
 #endif
-
-#define XATTR_DOSATTR_NAME "user.DOSATTRIB"
-#define XATTR_ATTRIBS_MASK (READ_ONLY_FILE | HIDDEN_FILE | SYSTEM_FILE | \
-  ARCHIVE_NEEDED)
-
-static int do_extr_xattr(const char *xbuf, ssize_t size, const char *name)
-{
-  if (size == -1 && errno == ENOTSUP) {
-    error("MFS: failed to get xattrs for %s, unsupported!\n", name);
-    return -1;
-  }
-  if (size <= 2 || strncmp(xbuf, "0x", 2) != 0)
-    return -1;
-  return strtol(xbuf + 2, NULL, 16) & XATTR_ATTRIBS_MASK;
-}
-
-static int get_dos_xattr(const char *fname)
-{
-  char xbuf[16];
-  ssize_t size = getxattr(fname, XATTR_DOSATTR_NAME, xbuf, sizeof(xbuf) - 1);
-  /* some dosemus forgot \0 so we fix it up here */
-  if (size > 0 && xbuf[size - 1] != '\0') {
-    error("MFS: fixup xattr for %s\n", fname);
-    xbuf[size++] = '\0';
-    setxattr(fname, XATTR_DOSATTR_NAME, xbuf, size, XATTR_REPLACE);
-  }
-  return do_extr_xattr(xbuf, size, fname);
-}
-
-static int get_dos_xattr_fd(int fd, const char *name)
-{
-  char xbuf[16];
-  ssize_t size = fgetxattr(fd, XATTR_DOSATTR_NAME, xbuf, sizeof(xbuf) - 1);
-  /* some dosemus forgot \0 so we fix it up here */
-  if (size > 0 && xbuf[size - 1] != '\0') {
-    error("MFS: fixup xattr\n");
-    xbuf[size++] = '\0';
-    fsetxattr(fd, XATTR_DOSATTR_NAME, xbuf, size, XATTR_REPLACE);
-  }
-  return do_extr_xattr(xbuf, size, name);
-}
-
-static int xattr_str(char *xbuf, int xsize, int attr)
-{
-  int ret = snprintf(xbuf, xsize, "0x%x", attr & XATTR_ATTRIBS_MASK);
-  assert(ret > 0);
-  return (ret + 1);  // include '\0'
-}
-
-static int xattr_err(int err, const char *name)
-{
-  if (err) {
-    error("MFS: failed to set xattrs for %s: %s\n", name, strerror(errno));
-//    leavedos(5);
-//    return ACCESS_DENIED;
-  }
-  return 0;
-}
-
-static int set_dos_xattr(const char *fname, int attr)
-{
-  char xbuf[16];
-  int err = setxattr(fname, XATTR_DOSATTR_NAME, xbuf,
-      xattr_str(xbuf, sizeof(xbuf), attr), 0);
-  if (err) {
-    struct stat st;
-    err = stat(fname, &st);
-    if (err)
-      return FILE_NOT_FOUND;
-    if (!(st.st_mode & S_IWUSR)) {
-      err = chmod(fname, st.st_mode | S_IWUSR);
-      if (err)
-        return ACCESS_DENIED;
-      err = setxattr(fname, XATTR_DOSATTR_NAME, xbuf,
-            xattr_str(xbuf, sizeof(xbuf), attr), 0);
-    }
-  }
-  return xattr_err(err, fname);
-}
-
-static int set_dos_xattr_fd(int fd, int attr, const char *name)
-{
-  char xbuf[16];
-  return xattr_err(fsetxattr(fd, XATTR_DOSATTR_NAME, xbuf,
-      xattr_str(xbuf, sizeof(xbuf), attr), 0), name);
-}
 
 static int get_attr_simple(int mode)
 {
